# docs/swagger.yaml
basePath: /
definitions:
  ai.DeviceAuthResponse:
    properties:
      device_code:
        type: string
      expires_in:
        type: integer
      interval:
        type: integer
      user_code:
        type: string
      verification_uri:
        type: string
    type: object
  ai.GitHubAuthResult:
    properties:
      error:
        type: string
      status:
        type: string
      token:
        type: string
    type: object
  ai.GitHubUser:
    properties:
      login:
        type: string
      name:
        type: string
    type: object
...
  models.ModelDetail:
    properties:
      contextLimit:
        type: integer
      displayName:
        type: string
      id:
        type: string
      inputCostPerM:
        type: number
      outputCostPerM:
        type: number
    type: object
...
  models.ProviderInfo:
    properties:
      authType:
        type: string
      configured:
        type: boolean
      contextLimit:
        type: integer
      defaultModel:
        type: string
      id:
        type: string
      models:
        items:
          $ref: '#/definitions/models.ModelDetail'
        type: array
      name:
        type: string
    type: object
  models.ProviderTestResult:
    properties:
      error:
        type: string
      latencyMs:
        type: integer
      ok:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the API for the Pad Analyzer project.
  title: Pad Analyzer API
  version: "1.0"
paths:
...
  /api/providers/copilot/poll:
    post:
      consumes:
      - application/json
      description: Polls GitHub for the status of the Copilot device flow authentication.
      parameters:
      - description: Poll Copilot Auth Request
        in: body
        name: request
        required: true
        schema:
          properties:
            deviceCode:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ai.GitHubAuthResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Poll Copilot OAuth
      tags:
      - providers
  /api/providers/copilot/revoke:
    post:
      description: Revokes the GitHub Copilot authentication for the current user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Revoke Copilot OAuth
      tags:
      - providers
  /api/providers/copilot/start:
    post:
      description: Initiates the GitHub Copilot device flow authentication.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ai.DeviceAuthResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Start Copilot OAuth
      tags:
      - providers
  /api/providers/copilot/user:
    get:
      description: Returns information about the currently authenticated Copilot user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ai.GitHubUser'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Copilot user info
      tags:
      - providers
  /api/providers/github/poll:
    post:
      consumes:
      - application/json
      description: Polls GitHub for the status of the device flow authentication.
      parameters:
      - description: Poll Auth Request
        in: body
        name: request
        required: true
        schema:
          properties:
            deviceCode:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ai.GitHubAuthResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Poll GitHub OAuth
      tags:
      - providers
  /api/providers/github/revoke:
    post:
      description: Revokes the GitHub authentication for the current user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Revoke GitHub OAuth
      tags:
      - providers
  /api/providers/github/start:
    post:
      description: Initiates the GitHub device flow authentication.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ai.DeviceAuthResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Start GitHub OAuth
      tags:
      - providers
  /api/providers/github/user:
    get:
      description: Returns information about the currently authenticated GitHub user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ai.GitHubUser'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get GitHub user info
      tags:
      - providers
  /api/providers/list:
    get:
      description: Returns a list of supported AI providers and their current status.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ProviderInfo'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List AI providers
      tags:
      - providers
  /api/providers/test:
    post:
      consumes:
      - application/json
      description: Tests the connection to a specific AI provider.
      parameters:
      - description: Test Connection Request
        in: body
        name: request
        required: true
        schema:
          properties:
            provider:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProviderTestResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Test provider connection
      tags:
      - providers
...
swagger: "2.0"